\name{fragebogen}
\alias{fragebogen}
\title{Kopfumfang und andere Merkmale der Studiengaenge GD und AVM!}
\usage{
fragebogen
}
\description{
Data frame mit den Kopfumfaengen und weiteren Merkmalen:\cr
- [alter] Alter -> num 16 - 99 (Jahre)\cr
- [geschlecht] Geschlecht -> num 1 (weiblich), 2 (männlich), 3 (divers)\cr
- [note_mathe] Letzte Mathenote -> num 1 - 6 (Note)\cr
- [note_annahme] Annahme der Note im Kurs Technical Applications & Data Analytics -> num 1 - 6 (Note)\cr
- [kopf] Kopfumfang -> num 40 - 70 (cm)\cr
- [angst] Angst vor dem Kurs -> num 1 - 6 (Note)\cr
- [interesse] Interesse am Thema -> num 1 - 6 (Note)\cr
- [praxis] gewuenschter Praxisanteil in den Vorlesungen -> num 10 - 90 (Prozent)\cr
der Studiengaenge Gamedesign & Management und Audio-Visuelle Medien (5. Semester, 2021).
}
\examples{
print(fragebogen)
}
