\name{kirche}
\alias{kirche}
\title{Church exits in Germany from 2017 to 2020!}
\description{
Data frame contains church exits in Germany from 2017 to 2020 from\cr
https://de.statista.com/statistik/daten/studie/4052/umfrage/kirchenaustritte-in-deutschland-nach-konfessionen/
}
\usage{
koelsch
}
\examples{
print(kirche)
}
