\name{koelsch}
\alias{koelsch}
\title{Colgone brewed beer (Koelsch) consumption from 2017 to 2020!}
\description{
Data frame contains Colgone brewed beer (Koelsch) consumption from 2017 to 2020 from\cr
https://de.statista.com/statistik/daten/studie/172265/umfrage/haeufigkeit-konsum-von-koelsch/
}
\usage{
koelsch
}
\examples{
print(koelsch)
}
