\name{kopfumfang}
\alias{kopfumfang}
\title{Kopfumfang der Studiengänge GD und AVM!}
\usage{
kopfumfang
}
\description{
Data frame mit den Kopfumfängen und weiteren Merkmalen\cr
- Alter [alter] -> num 16 - 99 (Jahre)\cr
- Geschlecht [geschlecht] -> num 1 (weiblich), 2 (männlich), 3 (divers)\cr
- Note Mathe [note_mathe] -> num 1 - 6 (Note)\cr
- Annahme der Note im Kurs Technical Applications & Data Analytics [note_annahme] -> num 1 - 6 (Note)\cr
- Angst vor dem Kurs [angst] -> num 1 - 6 (Note)\cr
- Interesse am Thema [interesse] -> num 1 - 6 (Note)\cr
- gewünschter Praxisanteil [praxis] -> num 10 - 90 (Prozent)\cr
der Studiengänge Gamedesign & Management und AudioVisuelle Medien (5. Semester, 2021).
}
\examples{
kopfumfang
}
