% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{conjuncts two data frames}
\usage{
combine(id, foreign_key, list1, list2, ...)
}
\arguments{
\item{id}{ID of the left list.}

\item{foreign_key}{Foreign key of the right list.}

\item{list1}{Numeric values of the left list.}

\item{list2}{Numeric values of the right list.}

\item{col1}{Optionally, a name of id / foreign key column / variable.}

\item{col2}{Optionally, a name of list1 column / variable.}

\item{col3}{Optionally, a name of the list2 column / variable.}
}
\value{
Conjuncted data frame.
}
\description{
`combine` conjuncts two data frames. In each data frame must be an identifier (id and foreign_key) and two lists containing numeric values.
}
\examples{
library(ourdata)

# combines two data frames
combined_list <- combine(imr$name, hdi$country, imr$value, hdi$hdi)

## With column names spezified
combined_list <- combine(imr$name, hdi$country, imr$value, hdi$hdi, col1 = "Country", col2 = "IMR", col3 = "HDI")
}
