\name{fragebogen}
\alias{fragebogen}
\title{Data from a survey of GD and AVM courses, contains the head circumference and other features!}
\description{
Data frame with the head circumference and other features, containing\cr
- [alter] Age -> num 16 - 99 (years)\cr
- [geschlecht] Sex -> num 1 (female), 2 (male), 3 (divers)\cr
- [note_mathe] Last math grade -> num 1 - 6 (German grading system)\cr
- [note_annahme] Expected grade in the Technical Applications & Data Analytics course -> num 1 - 6 (German grading system)\cr
- [kopf] head circumference -> num 40 - 70 (cm)\cr
- [angst] Fear of the course -> num 1 - 6 (1 high, 6 low)\cr
- [interesse] Interest in the topic -> num 1 - 6 (1 high, 6 low)\cr
- [praxis] Desired practical part in the lectures -> num 10 - 90 (percent)\cr
from a survey in the Game Design & Management and Audio-Visuelle Medien courses (5th semester, 2021)\cr
at Hochschule Fresenius University of Applied Sciences in Cologne.
}
\usage{
fragebogen
}
\examples{
print(fragebogen)
}
