\name{kirche}
\alias{kirche}
\title{Church exits in Germany from 2017 to 2020!}
\description{
Data frame contains church exits in Germany from 2017 to 2020 from
}
\usage{
kirche
}
\source{
The data was downloaded from: https://de.statista.com/statistik/daten/studie/4052/umfrage/kirchenaustritte-in-deutschland-nach-konfessionen/
}
\examples{
library(ourdata)

print(kirche)
}
