\name{kopfumfang}
\alias{kopfumfang}
\title{Kopfumfang der Studiengänge GD und AVM!}
\usage{
kopfumfang
}
\description{
Data frame mit den Kopfumfängen und weiteren Merkmalen:\cr
- [alter] Alter -> num 16 - 99 (Jahre)\cr
- [geschlecht] Geschlecht -> num 1 (weiblich), 2 (männlich), 3 (divers)\cr
- [note_mathe] Letzte Mathenote -> num 1 - 6 (Note)\cr
- [note_annahme] Annahme der Note im Kurs Technical Applications & Data Analytics -> num 1 - 6 (Note)\cr
- [kopf] Kopfumfang -> num 40 - 70 (cm)\cr
- [angst] Angst vor dem Kurs -> num 1 - 6 (Note)\cr
- [interesse] Interesse am Thema -> num 1 - 6 (Note)\cr
- [praxis] gewünschter Praxisanteil in den Vorlesungen -> num 10 - 90 (Prozent)\cr
der Studiengänge Gamedesign & Management und Audio-Visuelle Medien (5. Semester, 2021).
}
\examples{
kopfumfang
}
