\name{ourdata}
\alias{ourdata}
\title{Hello students!}
\usage{
help(ourdata)
}
\description{
Data Frames des Kurses Technical Applications & Data Analytics:\cr
- [imr] infant mortality rates\cr
- [hdi] human development index\cr
- [koelsch] Kölschkonsum 2017 bis 2020\cr
- [kirche] Kirchaustritte 2017 bis 2020\cr
- [kopfumfang] Kopfümfänge und andere Merkmal aus Studiengängen GD und AVM\cr
}
\examples{
help(ourdata)\cr
print(imr)\cr
print(hdi)\cr
print(koelsch)\cr
print(kirche)\cr
print(kopfumfang)
}
