\name{ourdata}
\alias{ourdata}
\title{Hello students!}
\usage{
help(ourdata)
}
\description{
Data Frames des Kurses Technical Applications & Data Analytics:\cr
- [imr] infant mortality rates\cr
- [hdi] human development index\cr
- [koelsch] Koelschkonsum 2017 bis 2020\cr
- [kirche] Kirchaustritte 2017 bis 2020\cr
- [kopfumfang] Kopfumfang und andere Merkmale aus Studiengaengen GD und AVM
}
\examples{
print(imr)
print(hdi)
print(koelsch)
print(kirche)
print(kopfumfang)
}
