\name{ourdata}
\alias{ourdata}
\title{Hello Students!}
\description{
Data frames of the Technical Applications & Data Analytics course, containing\cr
- [imr] Infant mortality rates\cr
- [hdi] Human development index\cr
- [koelsch] Colgone brewed beer (Koelsch) consumption from 2017 to 2020\cr
- [kirche] Church exits in Germany from 2017 to 2020\cr
- [fragebogen] Head circumference and other featrues of the GD und AVM courses\cr\cr
Functions of the course, containing\cr
- [ourddata()] Prints a welcome message.
- [combine(id, foreign_key, list1, list2, ...)] Combines two data frames.
}
\usage{
\code{help(ourdata)
}
}
\examples{
## Data frames
print(imr)
print(hdi)
print(koelsch)
print(kirche)
print(fragebogen)

## Functions
help(combine)
}
